+++ snuba/datasets/cdc/groupedmessage_entity.py
@@ -8,7 +8,7 @@ from snuba.datasets.storages.factory import get_cdc_storage
 from snuba.pipeline.simple_pipeline import SimplePipelineBuilder
 from snuba.query.data_source.join import JoinRelationship, JoinType
 from snuba.query.processors import QueryProcessor
-from snuba.query.processors.basic_functions import BasicFunctionsProcessor
+from snuba.query.processors.logical.basic_functions import BasicFunctionsProcessor
 from snuba.query.processors.object_id_rate_limiter import ProjectRateLimiterProcessor
