+++ snuba/datasets/storages/discover.py
@@ -16,9 +16,6 @@ from snuba.datasets.storages import StorageKey
 from snuba.datasets.storages.errors import storage as error_storage
 from snuba.datasets.storages.errors_common import mandatory_conditions
 from snuba.datasets.storages.transactions import storage as transactions_storage
-from snuba.query.processors.arrayjoin_keyvalue_optimizer import (
-    ArrayJoinKeyValueOptimizer,
-)
 from snuba.query.processors.empty_tag_condition_processor import (
     EmptyTagConditionProcessor,
 )
@@ -26,6 +23,9 @@ from snuba.query.processors.events_bool_contexts import EventsBooleanContextsPro
 from snuba.query.processors.mapping_optimizer import MappingOptimizer
 from snuba.query.processors.mapping_promoter import MappingColumnPromoter
 from snuba.query.processors.null_column_caster import NullColumnCaster
+from snuba.query.processors.physical.arrayjoin_keyvalue_optimizer import (
+    ArrayJoinKeyValueOptimizer,
+)
 from snuba.query.processors.physical.type_converters.hexint_column_processor import (
     HexIntColumnProcessor,
 )
