+++ snuba/datasets/storages/errors_common.py
@@ -11,15 +11,15 @@ from snuba.clickhouse.columns import SchemaModifiers as Modifiers
 from snuba.clickhouse.columns import String, UInt
 from snuba.query.conditions import ConditionFunctions, binary_condition
 from snuba.query.expressions import Column, Literal
-from snuba.query.processors.arrayjoin_keyvalue_optimizer import (
-    ArrayJoinKeyValueOptimizer,
-)
 from snuba.query.processors.empty_tag_condition_processor import (
     EmptyTagConditionProcessor,
 )
 from snuba.query.processors.events_bool_contexts import EventsBooleanContextsProcessor
 from snuba.query.processors.mapping_optimizer import MappingOptimizer
 from snuba.query.processors.mapping_promoter import MappingColumnPromoter
+from snuba.query.processors.physical.arrayjoin_keyvalue_optimizer import (
+    ArrayJoinKeyValueOptimizer,
+)
 from snuba.query.processors.physical.type_converters.hexint_column_processor import (
     HexIntColumnProcessor,
 )
