+++ snuba/datasets/storages/transactions_common.py
@@ -10,12 +10,6 @@ from snuba.clickhouse.columns import (
 )
 from snuba.clickhouse.columns import SchemaModifiers as Modifiers
 from snuba.clickhouse.columns import String, UInt
-from snuba.query.processors.array_has_optimizer import ArrayHasOptimizer
-from snuba.query.processors.arrayjoin_keyvalue_optimizer import (
-    ArrayJoinKeyValueOptimizer,
-)
-from snuba.query.processors.arrayjoin_optimizer import ArrayJoinOptimizer
-from snuba.query.processors.bloom_filter_optimizer import BloomFilterOptimizer
 from snuba.query.processors.conditions_enforcer import ProjectIdEnforcer
 from snuba.query.processors.empty_tag_condition_processor import (
     EmptyTagConditionProcessor,
@@ -23,6 +17,12 @@ from snuba.query.processors.empty_tag_condition_processor import (
 from snuba.query.processors.events_bool_contexts import EventsBooleanContextsProcessor
 from snuba.query.processors.mapping_optimizer import MappingOptimizer
 from snuba.query.processors.mapping_promoter import MappingColumnPromoter
+from snuba.query.processors.physical.array_has_optimizer import ArrayHasOptimizer
+from snuba.query.processors.physical.arrayjoin_keyvalue_optimizer import (
+    ArrayJoinKeyValueOptimizer,
+)
+from snuba.query.processors.physical.arrayjoin_optimizer import ArrayJoinOptimizer
+from snuba.query.processors.physical.bloom_filter_optimizer import BloomFilterOptimizer
 from snuba.query.processors.physical.type_converters.hexint_column_processor import (
     HexIntArrayColumnProcessor,
     HexIntColumnProcessor,
